Added git commit
The issue with adding the remote location and pushing to it is that git won't do that. There must be something to push, and cargo doesn't commit anything (thankfully), so the user must commit something before calling `git push`.

It seems mostly reasonable in this specific situation in this context to commit everything, although this, in general, isn't a great thing to do.